backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.1.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
et-xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-core==2.33.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
redis==5.0.1
aioredis==2.0.1
requests==2.32.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0